backend/requirements.txt
# Core API
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.1

# Elasticsearch
elasticsearch==8.17.0

# Redis
redis==5.2.1

# Kafka
confluent-kafka==2.6.1

# ETL / Data
pandas==2.2.3

# Observability
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-redis==0.50b0
opentelemetry-instrumentation-elasticsearch==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0

# Testing
pytest==8.3.4
httpx==0.28.1
pytest-asyncio==0.24.0
